app/api/views.py
#from django.shortcuts import render

from rest_framework import generics
from api.models import News 
from api.serializers import NewsSerializer

class NewsList(generics.ListCreateAPIView):
    queryset = News.objects.all()
    serializer_class = NewsSerializer

class NewsDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = News.objects.all()
    serializer_class = NewsSerializer
