app/requirements.txt
psycopg2==2.9.6
Pillow==9.5.0
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
pytz==2023.3
sqlparse==0.4.4
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
